Improve conventions class name
