+++ codegen/java/src/main/kotlin/io/spine/protodata/codegen/java/JavaValueConverter.kt
@@ -39,7 +39,7 @@ import io.spine.tools.code.Java
  */
 @Suppress("TooManyFunctions")
 public class JavaValueConverter(
-    private val typeConverter: JavaImplConvention
+    private val typeConverter: MessageOrEnumConvention
 ) : ValueConverter<Expression, Java>() {
 
     override fun toNull(type: Type): Expression = Null
